error解消
