+++ app/assets/stylesheets/homes.scss
@@ -1,0 +1,3 @@
+// Place all the styles related to the homes controller here.
+// They will automatically be included in application.css.
+// You can use Sass (SCSS) here: https://sass-lang.com/
