--- app/controllers/customer/items_controller.rb
@@ -1,2 +1,0 @@
-class Public::ItemsController < ApplicationController
-end
