+++ app/controllers/homes_controller.rb
@@ -1,0 +1,4 @@
+class HomesController < ApplicationController
+  def top
+  end
+end
