+++ app/helpers/homes_helper.rb
@@ -1,0 +1,2 @@
+module HomesHelper
+end
