+++ app/views/homes/top.html.erb
@@ -1,0 +1,2 @@
+<h1>Homes#top</h1>
+<p>Find me in app/views/homes/top.html.erb</p>
