+++ config/routes.rb
@@ -1,12 +1,23 @@
 Rails.application.routes.draw do
 
   
-  root to: 'public/homes#top'
+ 
+  root to: 'homes#top'
   get "homes/about" => "homes#about", as: "about"
-  devise_for :admins
-  devise_for :customers
   
-  
+
+# 顧客用
+# URL /customers/sign_in ...
+devise_for :customers,skip: [:passwords], controllers: {
+  registrations: "public/registrations",
+  sessions: 'public/sessions'
+}
+
+# 管理者用
+# URL /admin/sign_in ...
+devise_for :admin, skip: [:registrations, :passwords] ,controllers: {
+  sessions: "admin/sessions"
+}
  
  #顧客 
   scope module: :public do
@@ -17,7 +28,7 @@ Rails.application.routes.draw do
   
   resources :orders, only: [:new, :create, :index, :show]
     post "orders/confirm" => "orders#confirm"
-    get "orders/thanks" => "orders#thanks"
+    get "orders/complete" => "orders#complete"
  
   
   resources :addresses, only: [:create, :index, :edit, :update, :destroy]
