+++ test/controllers/homes_controller_test.rb
@@ -1,0 +1,8 @@
+require "test_helper"
+
+class HomesControllerTest < ActionDispatch::IntegrationTest
+  test "should get top" do
+    get homes_top_url
+    assert_response :success
+  end
+end
